refactor: App.tsx を HeaderBar + ResultsPane に分割
